{{cookiecutter.repo_name}}/requirements-test.txt
django>=1.5.1
coverage
coveralls
mock>=1.0.1
pytest-django>=2.6.2
flake8>=2.1.0
tox>=1.7.0
